Client max request count updated.

+++ Common/Entities/Clients/IOClientsEntity.cs
@@ -21,8 +21,8 @@ namespace IOBootstrap.NET.Common.Entities.Clients
         public string ClientDescription { get; set; }
 
         public int IsEnabled { get; set; }
-        public int RequestCount { get; set;  }
-        public int MaxRequestCount { get; set; }
+        public long RequestCount { get; set;  }
+        public long MaxRequestCount { get; set; }
 
         #endregion
